A dynamic recompiler block's validity is now separate from its starting code address, should fix most of the known recompiler fatals; The CD-ROM buffer is now dynamically malloc'd; Did a lot of bugfixing in cdrom.c and scsi_disk.c, Debian 8's modprobe aha1542 and the NT 3.10.404.1 Trantor 310b driver now work; Fixed the IDE IRQ/MIRQ mess and ATAPI DMA transfer length; Fixed the operation of mmutranslate() a bit; Commented out several instances of excess logging; Applied the mainline PCem commit prefetch commit; Readded missing Sound Blaster DSP MIDI behavior and gave it the correct MPU struct; All non-IBM 486 machines now correctly have the VLB flag set.

+++ src/scsi/scsi_x54x.h
@@ -416,7 +416,7 @@ typedef struct {
 		ATBusSpeed;
 
     char	*fw_rev;	/* The 4 bytes of the revision command information + 2 extra bytes for BusLogic */
-    uint8_t	bus;		/* Basically a copy of device flags */
+    uint16_t	bus;		/* Basically a copy of device flags */
     uint8_t	setup_info_len;
     uint8_t	max_id;
     uint8_t	pci_slot;
